+++ ng/samples/src/app/practices-ui-clarity-samples/date-picker/date-picker-view.component.ts
@@ -4,6 +4,7 @@ import { FormControl, FormGroup, Validators } from '@angular/forms';
 @Component({
     selector: 'app-date-picker-view',
     templateUrl: './date-picker-view.component.html',
+    standalone: false
 })
 export class DatePickerViewComponent {
     formGroup = new FormGroup({