+++ ng/samples/src/app/practices-ui-samples/list/list-view-source/list-view-source.component.ts
@@ -8,6 +8,7 @@ import { listEndpointMock } from '../endpointMock';
 @Component({
     selector: 'app-list-view-source',
     templateUrl: './list-view-source.component.html',
+    standalone: false
 })
 export class ListViewSourceComponent implements OnDestroy {
     lastRequestedApiUrl: string;