+++ ng/samples/src/app/practices-ui-samples/list/list-view/list-view.component.ts
@@ -3,5 +3,6 @@ import { Component } from '@angular/core';
 @Component({
     selector: 'app-list-view',
     templateUrl: './list-view.component.html',
+    standalone: false
 })
 export class ListViewComponent {}